7.2.4 Добавляем компонент статистики
