--- src/view/statistics.js
@@ -1,116 +1,0 @@
-import flatpickr from "flatpickr";
-import SmartView from "./smart.js";
-import {getCurrentDate} from "../utils/trip-event.js";
-
-const DAYS_TO_FULLWEEK = 6;
-
-const createStatisticsTemplate = (data) => {
-  const {tasks, dateFrom, dateTo} = data;
-  // const completedTripEventCount = countCompletedTaskInDateRange(tasks, dateFrom, dateTo);
-
-  return `<section class="statistics">
-          <h2 class="visually-hidden">Trip statistics</h2>
-
-          <div class="statistics__item statistics__item--money">
-            <canvas class="statistics__chart  statistics__chart--money" width="900"></canvas>
-          </div>
-
-          <div class="statistics__item statistics__item--transport">
-            <canvas class="statistics__chart  statistics__chart--transport" width="900"></canvas>
-          </div>
-
-          <div class="statistics__item statistics__item--time-spend">
-            <canvas class="statistics__chart  statistics__chart--time" width="900"></canvas>
-          </div>
-        </section>`;
-};
-
-export default class Statistics extends SmartView {
-  constructor(tasks) {
-    super();
-
-    this._data = {
-      tasks,
-      // По условиям техзадания по умолчанию интервал - неделя от текущей даты
-      dateFrom: (() => {
-        const date = getCurrentDate();
-        date.setDate(date.getDate() - DAYS_TO_FULLWEEK);
-        return date;
-      })(),
-      dateTo: getCurrentDate()
-    };
-
-    this._colorsCart = null;
-    this._daysChart = null;
-
-    this._dateChangeHandler = this._dateChangeHandler.bind(this);
-
-    this._setCharts();
-    this._setDatepicker();
-  }
-
-  removeElement() {
-    super.removeElement();
-
-    if (this._colorsCart !== null || this._daysChart !== null) {
-      this._colorsCart = null;
-      this._daysChart = null;
-    }
-
-    if (this._datepicker) {
-      this._datepicker.destroy();
-      this._datepicker = null;
-    }
-  }
-
-  getTemplate() {
-    return createStatisticsTemplate(this._data);
-  }
-
-  restoreHandlers() {
-    this._setCharts();
-    this._setDatepicker();
-  }
-
-  _dateChangeHandler([dateFrom, dateTo]) {
-    if (!dateFrom || !dateTo) {
-      return;
-    }
-
-    this.updateData({
-      dateFrom,
-      dateTo
-    });
-  }
-
-  _setDatepicker() {
-    if (this._datepicker) {
-      this._datepicker.destroy();
-      this._datepicker = null;
-    }
-
-    this._datepicker = flatpickr(
-        this.getElement().querySelector(`.statistic__period-input`),
-        {
-          mode: `range`,
-          dateFormat: `j F`,
-          defaultDate: [this._data.dateFrom, this._data.dateTo],
-          onChange: this._dateChangeHandler
-        }
-    );
-  }
-
-  _setCharts() {
-    if (this._colorsCart !== null || this._daysChart !== null) {
-      this._colorsCart = null;
-      this._daysChart = null;
-    }
-
-    const {tasks, dateFrom, dateTo} = this._data;
-    const colorsCtx = this.getElement().querySelector(`.statistic__colors`);
-    const daysCtx = this.getElement().querySelector(`.statistic__days`);
-
-    this._colorsCart = renderColorsChart(colorsCtx, tasks);
-    this._daysChart = renderDaysChart(daysCtx, tasks, dateFrom, dateTo);
-  }
-}
